+++ packages/logs/src/index.ts
@@ -1,3 +1,3 @@
-export { Logger, LogsMessage } from './domain/logger'
-export { LogsUserConfiguration, LoggerConfiguration, LogsGlobal, datadogLogs } from './boot/logs.entry'
+export { Logger, LogsMessage, StatusType, HandlerType } from './domain/logger'
+export { LogsUserConfiguration, LoggerConfiguration, LogsPublicApi as LogsGlobal, datadogLogs } from './boot/logs.entry'
 export { LogsEvent } from './logsEvent.types'